+++ app/helpers/users/inquiries_helper.rb
@@ -1,0 +1,2 @@
+module Users::InquiriesHelper
+end
