+++ app/views/users/inquiries/inquiries.html.erb
