+++ config/routes.rb
@@ -19,9 +19,9 @@ Rails.application.routes.draw do
 
   scope module: :users do
     root "homes#top"
-    get "inquiries" => "homes#inquiries", as: "inquiries"
-    post "inquiries" => "homes#mail"
-    get "done" => "homes#done", as: "done"
+    get "inquiries" => "inquiries#inquiries", as: "inquiries"
+    post "inquiries" => "inquiries#mail"
+    get "done" => "inquiries#done", as: "done"
     get "users/mypage" => "users#show", as: "mypage"
     get "users/leave" => "users#leave"
     patch "users/withdraw" => "users#withdraw"
